implement handling of selected med card on interaction button click
