+++ fit-as-a-fiddle/app/interactions/components/selection-section/SelectionSection.jsx
@@ -1,54 +1,79 @@
-'use client'
+"use client";
 
-import React, { useState } from 'react';
-import SelectedMeds from './SelectedMeds';
-import SearchBar from '@/app/components/SearchBar';
-import { Container, Button, Row, Col } from 'react-bootstrap';
+import React, { useState } from "react";
+import SelectedMeds from "./SelectedMeds";
+import SearchBar from "@/app/components/SearchBar";
+import { Container, Button, Row, Col } from "react-bootstrap";
 import { FaPlus } from "react-icons/fa";
 import styles from "./SelectionSection.module.css";
 
-/*
-  This is a client component for the MedicationCard component. 
-  Since components are server components by default, we need to explicitly 
-  define client components-- this enables us to use things like useState. 
-  All children of client components become part of this client bundle.
-  See Next.js App Router docs for more info.
-*/
+// Test data
+const selectedCards = [
+  {
+    drugName: "simvastatin 40 MG Oral Tablet [Zocor]",
+    rxcui: "152923",
+  },
+  {
+    drugName: "fluconazole 50 MG Oral Tablet [Diflucan]",
+    rxcui: "207106",
+  },
+  {
+    drugName: "bosentan 125 MG Oral Tablet",
+    rxcui: "656659",
+  },
+];
 
-// I don't think this is right. Just wanna toggle the deletion of the card. 
+const InformationText = () => (
+  <p className="mt-4 text-center">
+    Add two or more medications below to check for interactions.
+  </p>
+);
+
+// Renders content for the medication selection section of the Interactions page.
+const SelectionSection = () => {
+  const [isDeleted, setIsDeleted] = useState(false);
+  const [selectedMedCards, setSelectedMedCards] = useState(selectedCards);
+
+  // Prepares list of selectedMedCards for Interaction API request.
+  // The rxcui from each MedCard is saved into a string.
+  const convertSelectedMedsToString = () => {
+    const rxcuiList = selectedMedCards
+      .map((medication) => medication.rxcui)
+      .join("+");
+    console.log(rxcuiList);
+  };
+
+  // I don't think this is right. Just wanna toggle the deletion of the card.
 const handleDeleteCard = (index) => {
-  selectedCards.remove(index);
+
 };
 
-const InformationText = ({ text }) => (
-  <p className='mt-4 text-center'>{text}</p>
-);
+  // TODO: modify so api call is triggered on click
+  // should this be in its own .jsx file as a separate component?
+  const CheckInteractionsButton = () => (
+    <Row
+      className={`justify-content-end mt-3 mb-3 ${styles.checkInteractionsButton}`}
+    >
+      <Col xs="auto">
+        <Button variant="primary" onClick={convertSelectedMedsToString}>
+          {"Check interactions"}
+        </Button>
+      </Col>
+    </Row>
+  );
 
-// TODO: modify so api call is triggered on click
-const CheckInteractionsButton = () => (
-  <Row className={`justify-content-end mt-3 mb-3 ${styles.checkInteractionsButton}`}>
-    <Col xs="auto">
-      <Button variant="primary" onClick={"/* TRIGGER API CALL */"} >Check Interactions</Button>
-    </Col>
-  </Row>
-);
-
-const SelectionSection = () => {
-  const [isDeleted, setIsDeleted] = useState(false); // I have to make this a client component if I use useState(). HELP
-
+  // Renders medication search & selection section.
   return (
     <section>
-      <InformationText text="Add two or more medications below to check for interactions." />
+      <InformationText />
       <Container className={styles.container}>
         <div className={styles.search}>
-          <SearchBar
-            placeholder="Enter a drug name"
-            ButtonIcon={FaPlus}
-          />
+          <SearchBar placeholder="Enter a drug name" ButtonIcon={FaPlus} />
         </div>
 
         <SelectedMeds
-          onDelete={() => setIsDeleted(true)} // Is this even needed?
+          selectedCards={selectedCards}
+          onDelete={handleDeleteCard}
         />
 
         <CheckInteractionsButton />
